# shared-contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Property Management System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tenants": {
      "post": {
        "tags": [
          "tenant"
        ],
        "summary": "Create a new tenant (FR-TENANT-01, FR-TENANT-02)",
        "description": "Register a new tenant with encrypted ID card.\n\nThe ID card number is encrypted with Fernet before storage\nand is **never** returned in the response.\nAudit log is recorded automatically.",
        "operationId": "create_tenant_api_v1_tenants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tenants/search": {
      "get": {
        "tags": [
          "tenant"
        ],
        "summary": "Search tenants (FR-TENANT-04)",
        "description": "Search for tenants by name, phone, or email within a property.\n\nResponse follows SDD \u00a73.1 pagination format:\n``{\"data\": [...], \"meta\": {\"page\", \"limit\", \"total\", \"has_next\"}}``",
        "operationId": "search_tenants_api_v1_tenants_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scoping property UUID",
              "title": "Property Id"
            },
            "description": "Scoping property UUID"
          },
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "description": "Search query (min 3 chars)",
              "title": "Query"
            },
            "description": "Search query (min 3 chars)"
          },
          {
            "name": "search_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Field to search: name, phone, or email",
              "default": "name",
              "title": "Search By"
            },
            "description": "Field to search: name, phone, or email"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Tenants Api V1 Tenants Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTenantRequest": {
        "properties": {
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "full_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Full Name"
          },
          "id_card_number": {
            "type": "string",
            "maxLength": 13,
            "minLength": 13,
            "title": "Id Card Number"
          },
          "phone": {
            "type": "string",
            "maxLength": 10,
            "minLength": 10,
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "emergency_contact_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Name"
          },
          "emergency_contact_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Phone"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "property_id",
          "full_name",
          "id_card_number",
          "phone"
        ],
        "title": "CreateTenantRequest",
        "description": "Request body for POST /api/v1/tenants.\n\nValidation:\n    - ``phone`` must be a valid Thai phone number (10 digits, starts with 0).\n    - ``id_card_number`` must be 13 digits (Thai national ID format).\n    - ``email`` is optional but must be valid if provided."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TenantCreateResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TenantResponse"
          },
          "meta": {
            "type": "null",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "TenantCreateResponse",
        "description": "Wrapper for 201 CREATED response (SDD \u00a73.1)."
      },
      "TenantResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "emergency_contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Name"
          },
          "emergency_contact_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Phone"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "property_id",
          "full_name",
          "phone",
          "created_at"
        ],
        "title": "TenantResponse",
        "description": "Response body for a single Tenant resource (SDD \u00a73.1 format).\n\n**Security:** ``id_card_number_encrypted`` is deliberately omitted.\nSensitive personal data is never returned through the API."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}